A251 Move command metadata to commands.rs adjacent to implementations
CommandMeta (renamed from CommandEntry, no handler field) and each
ENTRY_* static now live in commands.rs next to the cmd_* methods they
describe. generate_help() and COMMANDS also move there. tcl_interp.rs
retains only the TCL binding layer: ShellContext, make_interp with an
explicit (entry, handler) pairing list, and the tcl_* glue handlers.
